f hmri_def.qMRI_maps.QA is set to 1 in the default file (default), the toolbox outputs a QualityAssessment matlab structure that provides objective measures of data quality. Contains the following fields : - ContrastCoreg: values of the coregistration parameters (translations and rotations) of the MTw and T1w to the PDw. Measure of inter-scan motion. - SDR2s: contains estimates of the standard deviation of the R2s maps calculated for each contrast (MTw, T1w and PDw) in white matter. Measure of intra-scan motion (Castella R et al., OHBM 2016; Arn L et al., OHBM 2016; Castella R et al., in prep.) - PD: mean and SD of the A map in white matter following bias-field correction. Primarily a measure of the efficacy of the bias-field correction i.e. of the level of bias in the PD map from the receive sensitivity profile.
If hmri_def.qMRI_maps.ACPCrealign is set to 1 in the default file (default), the toolbox modifies the header of the qMRI maps to realign the AC-PC line to MNI space using SPM’s ‘avg152T1’ canonical template. The coordinate transform is saved in a ACPCrealign.mat structure to allow the user to undo the transformation if necessary.