Added DICOM import to the hMRI menu
The tbx_scfg_hmri_dicom_import redirects to the hMRI DICOM import
implementation in hMRI-Toolbox/spm12 and sets the path to this
implementation as prioritary for all hMRI tools. This also holds for
spm_jsonread and spm_jsonwrite. Note that if you haven't started the
Batch first, these may not be used prioritarily.

This diversion trick allows you to keep the standard SPM12 DICOM
implementation untouched (button DICOM import, or in Batch > SPM >
Utils > Import > DICOM import).

We keep the metadata stuff isolated in spm12 directory since the ultimate
goal is to have it fully integrated into SPM...

Only briefly tested, but seems to do the job :).