Redirect config file for hMRI-Toolbox clean installation
The config file redirect towards the
full implementation of the toolbox (tbx_cfg_hmri) only if present in the Matlab
path. The toolbox can therefore be stored in a directory independent from the
SPM implementation and synch'd with the main hMRI-Toolbox repository whenever
needed. If tbx_cfg_hmri is not found in the Matlab path, the hMRI Tools are listed
in the SPM Batch GUI but not available. A brief help section provides the user
with instructions for hMRI Tools installation.