Removed warning for UNIX/MAC users
After some investigation (Tobias Leutritz) it appears that the problem is not UNIX related. The font size in the Batch GUI was responsible for the wrong rendering of the text. Changing the font size solved the problem.

NOTE: the font size can be changed in the Batch GUI menu (View>Change Font/Fontsize).