"Many Subjects" data specification method removed
Makes the code simpler and easier to maintain in the future.
Was not working anyway :/... The implementation was (failing at) converting the "Many Subjects" into "Few Subjects"
structure.
In practice, the "Few Subjects" method is used most of the time, and can easily be used to
implement loops over bigger groups of subjects.