Updated defaults (number of echoes for A)
Current default to calculate PD maps is TE=0 fit for PDw input, and
T2*-weighting bias correction. Number of averages for A, although not used
by default, has been set to 6 in case the fullOLS (TE=0 fit) were
disabled. With T2*-weighting bias correction, the bias is properly
corrected and the results benefits from a higher number of echoes (higher
SNR). The number 6 is somehow arbitrary. Should be adjusted to optimize
the SNR. If the number of acquired echoes is <6, this will be
automatically adjusted at runtime.