By default, no imperfect spoiling correction applied
To avoid confusion and resulting mistake (applying correction coefficients to the wrong sequence).
With the publication of "standard MPM protocols", it is likely to have new protocols implemented with sequences other than the customised sequences for which the correction coefficients have been calculated.
Therefore the TR and FA criteria are not sufficient any longer to identify the right set of correction coeficients to be used unambiguously :/...